Backend/auth-service/requirements.txt
# requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic-settings==2.7.0
PyJWT==2.10.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
email-validator==2.2.0

